requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
huggingface-hub==0.33.1
pydantic==2.11.7
# LangGraph and LangChain
langgraph==0.2.16
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-experimental==0.3.3
# Tools and utilities
requests==2.31.0
beautifulsoup4==4.12.2
duckduckgo-search==6.2.13
python-multipart==0.0.9
# Testing and CI/CD
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
# Monitoring and logging
structlog==24.4.0
